moved detector context
